hidden layer size cast to ints

+++ src/denoise_autoencoder.py
@@ -114,16 +114,16 @@ def mainFunc(argv):
     # plt.savefig('./ae_patching.png')
 
     print("Initializing model")
-    print("Input size: {}".format(conf.train_image_size*conf.train_image_size))
-    print("H1 size: {}".format(conf.train_image_size*conf.train_image_size/2))
-    print("H2 size: {}".format(conf.train_image_size*conf.train_image_size/2/2))
-    print("H3 size: {}".format(conf.train_image_size*conf.train_image_size/2/2/2))
-    print("H4 size: {}".format(conf.train_image_size*conf.train_image_size/2/2/2/2))
-    model = ae(n_input=conf.train_image_size*conf.train_image_size,
-               n_hidden_1=conf.train_image_size*conf.train_image_size/2,
-               n_hidden_2=conf.train_image_size*conf.train_image_size/2/2,
-               n_hidden_3=conf.train_image_size*conf.train_image_size/2/2/2,
-               n_hidden_4=conf.train_image_size*conf.train_image_size/2/2/2/2,
+    print("Input size: {}".format(int(conf.train_image_size*conf.train_image_size)))
+    print("H1 size: {}".format(int(conf.train_image_size*conf.train_image_size/2)))
+    print("H2 size: {}".format(int(conf.train_image_size*conf.train_image_size/2/2)))
+    print("H3 size: {}".format(int(conf.train_image_size*conf.train_image_size/2/2/2)))
+    print("H4 size: {}".format(int(conf.train_image_size*conf.train_image_size/2/2/2/2)))
+    model = ae(n_input=int(conf.train_image_size*conf.train_image_size),
+               n_hidden_1=int(conf.train_image_size*conf.train_image_size/2),
+               n_hidden_2=int(conf.train_image_size*conf.train_image_size/2/2),
+               n_hidden_3=int(conf.train_image_size*conf.train_image_size/2/2/2),
+               n_hidden_4=int(conf.train_image_size*conf.train_image_size/2/2/2/2),
                learning_rate=1e-3,
                dropout=conf.dropout_train)
